requirements.txt
python-telegram-bot==21.9
anthropic==0.49.0
aiohttp>=3.9.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
